Reference

The following summarizes the documentation of types and methods provided by the LocalFilters package. This information is also available from the REPL by typing ? followed by the name of a method or a type.

Index

LocalFilters.ConstantProducer

LocalFilters.Kernel

LocalFilters.Neighborhood

LocalFilters.RectangularBox

LocalFilters.BilateralFilter.bilateralfilter

LocalFilters.BilateralFilter.bilateralfilter!

LocalFilters.ball

LocalFilters.bottom_hat

LocalFilters.bottom_hat!

LocalFilters.closing

LocalFilters.closing!

LocalFilters.convolve

LocalFilters.convolve!

LocalFilters.default_start

LocalFilters.dilate

LocalFilters.dilate!

LocalFilters.erode

LocalFilters.erode!

LocalFilters.localextrema

LocalFilters.localextrema!

LocalFilters.localfilter

LocalFilters.localfilter!

LocalFilters.localmean

LocalFilters.localmean!

LocalFilters.opening

LocalFilters.opening!

LocalFilters.strel

LocalFilters.top_hat

LocalFilters.top_hat!

Linear filters

LocalFilters provides convolutive filters which are shift-invariant linear filters.

LocalFilters.localmean — Function

localmean(A, B=3)

yields the local mean of A in a neighborhood defined by B. The result is an array similar to A. If B is not specified, the neighborhood is a hyperrectangular moving window of size 3 in every dimension. Otherwise, B may be specified as a Cartesian box, or as an array of booleans of same number of dimensions as A. If B is a single odd integer (as it is by default), the neighborhood is assumed to be a hyperrectangular moving window of size B in every dimension.

See also localmean! and localfilter!.

LocalFilters.localmean! — Function

localmean!(dst, A, B=3) -> dst

overwrites dst with the local mean of A in a neighborhood defined by B and returns dst.

See also localmean and localfilter!.

LocalFilters.convolve — Function

convolve(A, B)

yields the discrete convolution of array A by the kernel defined by B. The result dst is an array similar to A.

Using Sup(A) to denote the set of valid indices for array A and assuming B is an array of values, the discrete convolution of A by B writes:

T = promote_type(eltype(A), eltype(B))
for i ∈ Sup(A)
    v = zero(T)
    @inbounds for k ∈ Sup(B) ∩ (i - Sup(A))
        v += A[i-k]*B[k]
    end
    dst[i] = v
end

with T the type of the product of elements of A and B, and where Sup(B) ∩ (i - Sup(A)) denotes the subset of indices k such that k ∈ Sup(B) and i - k ∈ Sup(A) and thus for which B[k] and A[i-k] are valid.

See also convolve! and localfilter!.

LocalFilters.convolve! — Function

convolve!(dst, A, B) -> dst

overwrites dst with the discrete convolution of A by the kernel B and returns dst.

See also convolve and localfilter!.

Mathematical morphology

LocalFilters.erode — Function

erode(A, R=3) -> Amin

yields the erosion of A by the structuring element defined by R. The erosion is the array of local minima of A. The returned result Amin is similar to A (same size and type).

If the structuring element R is a simple hyperrectangular moving window, the much faster van Herk-Gil-Werman algorithm is used. If specified as an odd integer (as is assumed by default), the structuring element is a hyperrectangular moving window of size R along every dimension of A.

An erosion is one of the most basic operations of mathematical morphology. See erode! for an in-place version of the method, dilate for retrieving the local maxima, and localextrema for performing an erosion and a dilation in a single pass.

LocalFilters.erode! — Function

erode!(Amin, A, R=3) -> Amin

overwrites Amin with the erosion of the array A by the structuring element defined by R and returns Amin.

If the structuring element R is a simple hyperrectangular moving window, the much faster van Herk-Gil-Werman algorithm is used and the operation can be done in-place. That is, A and Amin can be the same arrays. In that case, the following syntax is allowed:

erode!(A, R=3) -> A

See erode for an out-of-place version and for more information.

LocalFilters.dilate — Function

dilate(A, R=3) -> Amax

yields the dilation of A by the structuring element defined by R. The dilation is the array of local maxima of A. The returned result Amax is similar to A (same size and type).

If the structuring element R is a simple hyperrectangular moving window, the much faster van Herk-Gil-Werman algorithm is used. If specified as an odd integer (as is assumed by default), the structuring element is a hyperrectangular moving window of size R along every dimension of A.

A dilation is one of the most basic operations of mathematical morphology. See dilate! for an in-place version of the method, erode for retrieving the local minima, and localextrema for performing an erosion and a dilation in a single pass.

LocalFilters.dilate! — Function

dilate!(Amax, A, R=3) -> Amax

overwrites Amax with a dilation of the array A by the structuring element defined by R and returns Amax.

If the structuring element R is a simple hyperrectangular moving window, the much faster van Herk-Gil-Werman algorithm is used and the operation can be done in-place. That is, A and Amin can be the same arrays. In that case, the following syntax is allowed:

dilate!(A, R=3) -> A

See dilate for an out-of-place version and for more information.

LocalFilters.localextrema — Function

localextrema(A, R=3) -> Amin, Amax

yields the results of performing an erosion and a dilation of A by the structuring element defined by R in a single pass. Calling this method is usually almost twice as fast as calling erode and dilate.

See localextrema! for an in-place version of the method, and erode or dilate for a description of these operations.

LocalFilters.localextrema! — Function

localextrema!(Amin, Amax, A, R=3) -> Amin, Amax

overwrites Amin and Amax with, respectively, an erosion and a dilation of the array A by the structuring element defined by R in a single pass.

See localextrema for an out-of-place version for more information.

LocalFilters.closing — Function

closing(A, R=3) -> dst

yields a closing of array A by the structuring element defined by R. A closing is a dilation followed by an erosion. The result dst is an array similar to A.

See closing! for an in-place version of the method, opening for a related filter, and erode or dilate for a description of these operations.

LocalFilters.closing! — Function

closing!(dst, wrk, A, R=3) -> dst

overwrites dst with the result of a closing of A by the structuring element defined by R using wrk as a workspace array. The arguments dst, wrk, and A must be similar arrays, dst and A may be identical, but wrk must not be the same array as A or dst. The destination dst is returned.

See closing for a description of this kind of filter and for the meaning of the arguments.

LocalFilters.opening — Function

opening(A, R=3) -> dst

yields an opening of array A by the structuring element defined by R. An opening is an erosion followed by a dilation. The result dst is an array similar to A.

See opening! for an in-place version of the method, closing for a related filter, and erode or dilate for a description of these operations.

LocalFilters.opening! — Function

opening!(dst, wrk, A, R=3) -> dst

overwrites dst with the result of an opening of A by the structuring element defined by R using wrk as a workspace array. The arguments dst, wrk, and A must be similar arrays, dst and A may be identical, but wrk must not be the same array as A or dst. The destination dst is returned.

See opening for a description of this kind of filter and for the meaning of the arguments.

LocalFilters.top_hat — Function

top_hat(A, R=3 [, S]) -> dst

performs a summit detection by applying a top-hat filter to array A using the structuring element defined by R for the feature detection. Top-hat filtering is equivalent to:

dst = A .- opening(A, R)

Optional argument S specifies the structuring element for smoothing A prior to top-hat filtering. If R and S are specified as the radii of the structuring elements, then S should be smaller than R. For instance:

top_hat(bitmap, 3, 1)

may be used to detect text or lines in a bitmap image.

See bottom_hat for a related operation, LocalFilters.top_hat! for an in-place version.

LocalFilters.top_hat! — Function

LocalFilters.top_hat!(dst, wrk, A, R=3) -> dst

overwrites dst with the result of a top-hat filter applied to A with structuring element R, and using wrk as a workspace whose contents is not preserved. The arguments A, dst, and wrk must be similar but different arrays. The destination dst is returned.

See also top_hat for more details.

LocalFilters.bottom_hat — Function

bottom_hat(A, R=3 [, S]) -> dst

performs a valley detection by applying a bottom-hat filter to array A using the structuring element defined by R for the feature detection. Bottom-hat filtering is equivalent to:

dst = closing(A, R) .- A

Optional argument S specifies the structuring element for smoothing A prior to bottom-hat filtering. If R and S are specified as the radii of the structuring elements, then S should be smaller than R.

See top_hat for a related operation, LocalFilters.bottom_hat! for an in-place version.

LocalFilters.bottom_hat! — Function

LocalFilters.bottom_hat!(dst, wrk, A, R=3) -> dst

overwrites dst with the result of a bottom-hat filter applied to A with structuring element R and optional smoothing element S. Argument wrk is a workspace array whose contents is not preserved. The arguments A, dst, and wrk must be similar but different arrays. The destination dst is returned.

See also bottom_hat for more details.

LocalFilters.strel — Function

strel(T, A)

yields a structuring element suitable for mathematical morphology operations. The result is a Kernel whose elements have type T (which can be Bool or a floating-point type). Argument A can be a Cartesian box or a Kernel with boolean elements.

If T is a floating-point type, then the result is a so-called flat structuring element whose coefficients are zero(T) inside the shape defined by A and -T(Inf) elsewhere.

Other non-linear filters

LocalFilters.BilateralFilter.bilateralfilter — Function

bilateralfilter([T=float(eltype(A)),] A, F, G, ...)

yields the result of applying the bilateral filter on array A.

Argument F specifies how to smooth the differences in values. It may be function which takes two values from A as arguments and returns a nonnegative weight. It may be a real which is assumed to be the standard deviation of a Gaussian.

Arguments G, ... specify the settings of the distance filter for smoothing differences in coordinates. There are several possibilities:

G, ... can be a LocalFilters.Kernel instance (specified as a single argument).

Argument G may be a function taking as argument the Cartesian index of the coordinate differences and returning a nonnegative weight. Argument G may also be a real specifying the standard deviation of the Gaussian used to compute weights. Subsequent arguments ... are to specify the neighborhood where to apply the distance filter function, they can be a Neighborhood object such as a RectangularBox or anything that may defined a neighborhood such as an odd integer assumed to be the width of the neighborhood along every dimensions of A. If a standard deviation σ is specified for G with no subsequent arguments, a default window of size ±3σ is assumed.

Optional argument T can be used to force the element type used for (most) computations. This argument is needed if the element type of A is not a real.

See bilateralfilter! for an in-place version of this function and see Wikipedia for a description of the bilateral filter.

LocalFilters.BilateralFilter.bilateralfilter! — Function

bilateralfilter!([T=float(eltype(A)),] dst, A, F, G, ...) -> dst

overwrites dst with the result of applying the bilateral filter on array A and returns dst.

See bilateralfilter for a description of the other arguments than dst and see Wikipedia for a description of the bilateral filter.

Generic driver for custom local filters

LocalFilters.localfilter — Function

localfilter([T=eltype(A),] A, dims, op, rngs[, wrk]) -> dst

yields the result of applying van Herk-Gil-Werman algorithm to filter array A along dimension(s) dims with (associative) binary operation op and contiguous structuring element(s) defined by the interval(s) rngs. Optional argument wrk is a workspace array which is automatically allocated if not provided; otherwise, it must be a vector with the same element type as A and it is resized as needed (by calling the resize! method). The optional argument T allows to specify another type of element than eltype(A) for the result.

Argument dims specifies along which dimension(s) of A the filter is to be applied, it can be a single integer, a tuple of integers, or a colon : to apply the operation to all dimensions. Dimensions are processed in the order given by dims (the same dimension may appear several times) and there must be a matching interval in rngs to specify the structuring element (except that if rngs is a single interval, it is used for every dimension in dims). An interval is either an integer or an integer valued unit range in the form kmin:kmax (an interval specified as a single integer, say k, is the same as specifying k:k).

Assuming a mono-dimensional array A, the single filtering pass:

dst = localfilter(A, :, op, rng)

amount to computing:

dst[j] = A[j-kmax] ⋄ A[j-kmax+1] ⋄ A[j-kmax+2] ⋄ ... ⋄ A[j-kmin]

for all j ∈ [first(axes(A,1)):last(axes(A,1))], with x ⋄ y = op(x, y), kmin = first(rng) and kmax = last(rng). Note that if kmin = kmax = k (which occurs if rng is a simple integer), the result of the filter is to operate a simple shift by k along the corresponding dimension and has no effects if k = 0. This can be exploited to not filter some dimension(s).

The morphological erosion (local minimum) of the array A on a centered structuring element of width 7 in every dimension can be obtained by:

localfilter(A, :, min, -3:3)

Index interval 0:0 may be specified to do nothing along the corresponding dimension. For instance, assuming A is a three-dimensional array:

localfilter(A, :, max, (-3:3, 0:0, -4:4))

yields the morphological dilation (i.e. local maximum) of A in a centered local neighborhood of size 7×1×9 (nothing is done along the second dimension). The same result may be obtained with:

localfilter(A, (1,3), max, (-3:3, -4:4))

where the second dimension is omitted from the list of dimensions.

The local average of the two-dimensional array A on a centered moving window of size 11×11 can be computed as:

localfilter(A, :, +, (-5:5, -5:5))*(1/11)

See localfilter! for an in-place version of the method.

LocalFilters.localfilter! — Function

localfilter!([dst = A,] A, dims, op, rngs[, wrk])

overwrites the contents of dst with the result of applying van Herk-Gil-Werman algorithm to filter array A along dimension(s) dims with (associative) binary operation op and contiguous structuring element(s) defined by the interval(s) rngs and using optional argument wrk as a workspace array. The destination dst must have the same indices as the source A (that is, axes(dst) == axes(A)). Operation may be done in-place and dst and A can be the same; this is the default behavior if dst is not specified.

See localfilter for a full description of the method.

The in-place morphological erosion (local minimum) of the array A on a centered structuring element of width 7 in every dimension can be obtained by:

localfilter!(A, :, min, -3:3)

localfilter!(dst, A, B, initial, update, store!) -> dst

overwrites the destination dst with the result of a local filter applied to the source A, on a relative neighborhood defined by B, and implemented by the functions initial, update, and store!. The purpose of these functions is explained by the following pseudo-code implementing the local fitering:

@inbounds for i ∈ Sup(A)
    v = initial(A[i])
    for j ∈ Sup(A) ∩ (i - Sup(B))
        v = update(v, A[j], B[i-j])
    end
    store!(dst, i, v)
end

where Sup(A) denotes the support of A (that is the set of indices in A) and i - Sup(B) denotes the set of indices j such that i - j ∈ Sup(B) with Sup(B) the support of B. In other words, j ∈ Sup(A) ∩ (i - Sup(B)) means all indices j such that j ∈ Sup(A) and i - j ∈ Sup(B) so that A[j] and B[i-j] are in-bounds.

Warning

The loop(s) in localfilter! are performed without bounds checking of the destination and it is the caller's responsibility to insure that the destination have the correct size. It is however always possible to write store! so that it performs bounds checking.

For example, implementing a local minimum filter (that is, an erosion), is as simple as:

localfilter!(dst, A, B,
             (a)     -> typemax(a),
             (v,a,b) -> min(v,a),
             (d,i,v) -> @inbounds(d[i] = v))

As another example, implementing a convolution by B writes:

localfilter!(dst, A, B,
             (a)     -> zero(a),
             (v,a,b) -> v + a*b,
             (d,i,v) -> @inbounds(d[i] = v))

Note

If the methods init, update, and/or store! are anonymous functions or closures, beware that they should not depend on local variables because this may have a strong impact on performances. For instance, you may consider using LocalFilters.ConstantProducer as a replacement for init (have a look at the implementation of methods such as localmean! or convolve!).

Types

LocalFilters.Neighborhood — Type

Neighborhood

is the abstract type of which all neighborhood types inherit.

LocalFilters.RectangularBox — Type

RectangularBox(start, stop)

yields a neighborhood which is a rectangular (Cartesian) box defined by the bounds of the multi-dimensional indices in the box.

Another possibility is to specify the dimensions of the box and the offsets of its central element:

RectangularBox(dims, offs)

with dims a N-tuple of dimensions and offs either a N-tuple of indices of an instance of CartesianIndex{N}.

A RectangularBox can also be defined by the index ranges along all the dimensions. For example:

RectangularBox(-3:3, 0, -2:1)
RectangularBox((-3:3, 0, -2:1))

both yield a 3-dimensional RectangularBox of size 7×1×4 and whose first index varies on -3:3, its second index is 0 while its third index varies on -2:1.

Finally, a RectangularBox can be defined as:

RectangularBox(R)

where R is an instance of CartesianIndices.

LocalFilters.Kernel — Type

A Kernel can be used to define a weighted neighborhood (for weighted local average or for convolution) or a structuring element (for mathematical morphology). It is a rectangular array of coefficients over a, possibly off-centered, rectangular neighborhood. In general, it is sufficient to specify ::Kernel{T,N} in the signature of methods, with T the type of the coefficients and N the number of dimensions (the third parameter A of the type is to fully qualify the type of the array of coefficients).

A kernel is built as:

B = Kernel{T}(C, start=default_start(C))

where C is the array of coefficients (which can be retrieved by coefs(B)) and start the initial CartesianIndex for indexing the kernel (which can be retrieved by first_cartesian_index(B)). The start parameter let the caller choose an arbitrary origin for the kernel coefficients; when a filter is applied, the following mapping is assumed:

B[k] ≡ C[k + off]

where off = first_cartesian_index(C) - first_cartesian_index(B).

If start is omitted, its value is set so that the origin (whose index is zero(CartesianIndex{N}) with N the number of dimensions) of the kernel indices is at the geometric center of the array of coefficients (see LocalFilters.default_start). Optional type parameter T is to impose the type of the coefficients.

To convert the element type of the coefficients of an existing kernel, do:

Kernel{T}(K)

which yields a kernel whose coefficients are those of the kernel K converted to type T.

It is also possible to convert instances of RectangularBox into a kernel with boolean coefficients by calling:

Kernel(B)

where B is the neighborhood to convert into an instance of Kernel.

LocalFilters.ConstantProducer — Type

ConstantProducer(val)

yields a functor object which, when called as a function with any number of arguments of any types, always yields the value val.

This is useful to avoid closures which are quite inefficient and thus have a strong impact on performances when called repeatedly in loops. Typically when calling localfilter! with an initializer whose value is a constant but given by an anonymous function or a closure depending on a local variable, the code:

v0 = ... # initial state variable
localfilter!(dst, A, B, ConstantProducer(v0), update, store!)

is much faster than:

v0 = ... # initial state variable
localfilter!(dst, A, B, a -> v0, update, store!)

Utilities

LocalFilters.default_start — Function

default_start(A) -> I::CartesianIndex

yields the initial (multi-dimensional) index of a rectangular region which has the same size as the array A but whose origin (that is, index zero(CartesianIndex{N})) is at the geometrical center of the region (with the same conventions as fftshift.

LocalFilters.ball — Function

LocalFilters.ball(N, r)

yields a boolean mask which is a N-dimensional array with all dimensions odd and equal and set to true where position is inside a N-dimensional ball of radius r.